+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Add attributes.class method
 - Add `@blank?`, `@defined?`, `@empty?`, `@nil?` and `@present?`
 - Add `@method`, `@patch`, `@put` and `@delete`
+- Add `@pushif` and `@prependif`
 - Add string methods to slots
 - Fix whitespaces causing problems between case and when statements
 - Pass Rails `session`, `cookies`, `flash` and `params` variables into components